Add content_navigation graph for authored users
