+++ app/src/main/java/skarlat/dev/ecoproject/fragment/HomeViewNodel.kt
@@ -1,0 +1,5 @@
+package skarlat.dev.ecoproject.fragment
+
+class HomeViewModel: BaseViewModel() {
+
+}
