+++ app/src/main/res/layout/activity_home.xml
@@ -3,62 +3,17 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/home_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/colorWhite"
-    android:orientation="vertical"
-    tools:context=".activity.HomeActivity">
+    tools:viewBindingIgnore="true"
+    >
 
-    <LinearLayout
-        android:id="@+id/linearLayout"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/nav_host"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:padding="16dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-
-            <TextView
-                android:id="@+id/hello_user"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:fontFamily="@font/bellota_bold"
-                android:paddingRight="10dp"
-                android:textSize="24sp"
-                tools:text="Привет, Кристина!" />
-
-            <ImageView
-                android:id="@+id/profile_image"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:layout_centerHorizontal="true"
-                android:layout_weight="0"
-                android:src="@drawable/ic_profile_icon"
-                tools:src="@drawable/ic_profile_icon" />
-        </LinearLayout>
-
-        <TextView
-            android:id="@+id/welcome_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="2dp"
-            android:layout_marginRight="40dp"
-            android:text="Отличный день для того, чтобы узнать\nчто-то новое!" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recycle_courses"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="24dp"
-            android:clipChildren="false"
-            android:clipToPadding="false"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            tools:itemCount="1"
-            tools:listitem="@layout/card_course" />
-    </LinearLayout>
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/content_navigation"
+        />
 </FrameLayout>
