+++ app/src/main/res/navigation/content_navigation.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/content_navigation"
+    app:startDestination="@id/homeFragment"
+    >
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="skarlat.dev.ecoproject.fragment.HomeFragment"
+        android:label="HomeFragment"
+        tools:layout="@layout/fragment_home"
+        >
+        <action
+            android:id="@+id/action_homeFragment_to_courseActivity"
+            app:destination="@id/courseActivity"
+            />
+    </fragment>
+    <activity
+        android:id="@+id/courseActivity"
+        android:name="skarlat.dev.ecoproject.activity.CourseActivity"
+        android:label="activity_course_card"
+        tools:layout="@layout/activity_course_card"
+        />
+</navigation>
